requirements.txt
pandas
psycopg2-binary
pytest
python-dotenv
sentence-transformers
